refactor(audit): f-9 split ffmpeg_utils into discovery + decode modules
#115 finding 9: core/ffmpeg_utils.py (535 lines) conflated FFmpeg
binary discovery/install with TrueHD/MLP decode + read_audio. Split into:
- core/ffmpeg_discovery.py: get_ffmpeg_version, find/install/setup,
  ensure/check_ffmpeg_available, and the lazy-init FFMPEG_PATH/FFPROBE_PATH
  globals (the only mutable state).
- core/audio_truehd.py: is_truehd_file, convert, get_truehd_channel_info/
  profile, is_truehd_atmos_object_master, read_audio, get_supported_audio_formats.
  It reads the live paths via `_discovery.FFMPEG_PATH/FFPROBE_PATH` so lazy
  init still works across the module boundary.
- core/ffmpeg_utils.py: thin re-export shim, so core.utils and the gui_main
  smoke test keep importing it unchanged.

Pure code move (no logic edits). read_audio is on the BRIR input path: demo
md5 unchanged (07eef9ef…). The lazy-setup regression test is repointed at the
new owning modules (every assertion preserved); gui_main smoke list + the
CLAUDE.md architecture map gain the two modules. Nuitka --standalone follows
core.* imports automatically, so no build-flag change is needed.

Full suite 195 pass / 0 fail / 3 skip; ruff clean.
