+++ CLAUDE.md
@@ -22,7 +22,9 @@ core/
   microphone_deviation_correction.py  ← 마이크 편차 보정
   recorder.py             ← 녹음/재생
   utils.py                ← WAV I/O, DSP/플롯 헬퍼 (Phase 5에서 FFmpeg 분리)
-  ffmpeg_utils.py         ← FFmpeg 검색/설치 + TrueHD 변환 (Phase 5 분리)
+  ffmpeg_discovery.py     ← FFmpeg 검색/설치 + lazy 초기화 globals (audit #115-9 분리)
+  audio_truehd.py         ← TrueHD/MLP 디코드 + read_audio (audit #115-9 분리)
+  ffmpeg_utils.py         ← 위 둘의 하위 호환 re-export 셸
   constants.py            ← 스피커 이름/딜레이 등 상수
   pipeline.py             ← ProcessingConfig + BRIRPipeline (Phase 2)
   cli_builder.py          ← argparse 자동 생성기 (Phase 3)
