+++ gui_main.py
@@ -75,6 +75,8 @@ def _smoke_test():
         "core.plotting.hrir_plotter",
         "core.plotting.impulse_response_plotter",
         "core.ffmpeg_utils",
+        "core.ffmpeg_discovery",
+        "core.audio_truehd",
         "i18n.localization",
         "updater.update_checker",
         "updater.updater_core",
